backend/ai_service/requirements.txt
fastapi
uvicorn
torch
torchvision
faiss-cpu
pillow
numpy
python-multipart
requests
pymongo
python-dotenv
mediapipe
gradio_client
